Refactor tests to use renderWithThemeAndProviders from reduxRenderTestUtils
